dt is a float again and mem usage in debug

+++ vroom/time.py
@@ -1,3 +1,3 @@
 class Time:
     # delta time in milliseconds
-    dt: int = 0
+    dt: float = 0
